extract out the multi query functionality
